Scribo/Backend/requirements.txt
Flask==2.0.3
pytesseract==0.3.8
Pillow==8.4.0
transformers==4.11.3
torch==1.9.0
pymongo==3.12.1
nltk==3.6.5